Front: adding typing to profile
